+++ app/controllers/application_controller.rb
@@ -16,4 +16,12 @@ class ApplicationController < ActionController::Base
                :password, :password_confirmation
     end
   end
+
+  def scrub_order(model, params_order, default_order)
+    (params_column, params_direction) = params_order.to_s.strip.downcase.split(' ')
+    direction = (params_direction == 'desc' ? 'DESC' : nil)
+    column_name = (model.column_names.collect { |c| model.table_name + '.' + c }.find { |c| c == params_column })
+    order = column_name.blank? ? default_order : [column_name, direction].compact.join(' ')
+    order
+  end
 end